Refactor radial_distribution_between_species()
