+++ src/gemdat/plots/matplotlib/__init__.py
@@ -17,7 +17,6 @@ from ._jumps_vs_distance import jumps_vs_distance
 from ._jumps_vs_time import jumps_vs_time
 from ._msd_per_element import msd_per_element
 from ._radial_distribution import radial_distribution
-from ._radial_distribution_between_species import radial_distribution_between_species
 from ._rectilinear import rectilinear
 from ._shape import shape
 from ._vibrational_amplitudes import vibrational_amplitudes
@@ -37,7 +36,6 @@ __all__ = [
     'jumps_vs_time',
     'msd_per_element',
     'radial_distribution',
-    'radial_distribution_between_species',
     'rectilinear',
     'shape',
     'vibrational_amplitudes',
