+++ src/gemdat/plots/matplotlib/_radial_distribution.py
@@ -26,13 +26,14 @@ def radial_distribution(rdfs: Iterable[RDFData]) -> matplotlib.figure.Figure:
     fig, ax = plt.subplots()
 
     for rdf in rdfs:
-        ax.plot(rdf.x, rdf.y, label=rdf.symbol)
+        ax.plot(rdf.x, rdf.y, label=rdf.label)
 
-    states = ', '.join({rdf.state for rdf in rdfs})
+    states = ', '.join({rdf.state for rdf in rdfs if rdf.state})
+    state_suffix = f' ({states})' if states else ''
 
     ax.legend()
     ax.set(
-        title=f'Radial distribution function ({states})',
+        title=f'Radial distribution function{state_suffix}',
         xlabel='Distance (Å)',
         ylabel='Counts',
     )
